+++ modules/pytket-qiskit/_metadata.py
@@ -1,2 +1,2 @@
-__extension_version__ = "0.15.0"
+__extension_version__ = "0.16.1"
 __extension_name__ = "pytket-qiskit"