+++ modules/pytket-qiskit/tests/backend_test.py
@@ -15,13 +15,13 @@ import json
 import os
 import sys
 from collections import Counter
-from typing import Dict, Any, Tuple, cast
+from typing import Dict, Any, cast
 import math
 import cmath
 import pickle
 from hypothesis import given, strategies
 import numpy as np
-from pytket.circuit import Circuit, OpType, BasisOrder, Node, Qubit, reg_eq  # type: ignore
+from pytket.circuit import Circuit, OpType, BasisOrder, Qubit, reg_eq  # type: ignore
 from pytket.passes import CliffordSimp  # type: ignore
 from pytket.pauli import Pauli, QubitPauliString  # type: ignore
 from pytket.predicates import CompilationUnit, NoMidMeasurePredicate  # type: ignore
@@ -53,6 +53,7 @@ from qiskit import IBMQ  # type: ignore
 from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
 from qiskit.circuit import Parameter  # type: ignore
 from qiskit.providers.aer.noise.noise_model import NoiseModel  # type: ignore
+from qiskit.providers.aer.noise import ReadoutError  # type: ignore
 from qiskit.providers.aer.noise.errors import depolarizing_error, pauli_error  # type: ignore
 import pytest
 
@@ -78,7 +79,7 @@ def circuit_gen(measure: bool = False) -> Circuit:
     return c
 
 
-def get_test_circuit(measure: bool) -> QuantumRegister:
+def get_test_circuit(measure: bool) -> QuantumCircuit:
     qr = QuantumRegister(5)
     cr = ClassicalRegister(5)
     qc = QuantumCircuit(qr, cr)
@@ -150,7 +151,7 @@ def test_noise() -> None:
     c.measure_all()
     b = AerBackend(noise_model)
     n_shots = 50
-    b.compile_circuit(c)
+    c = b.get_compiled_circuit(c)
     shots = b.get_shots(c, n_shots, seed=4)
     zer_exp = []
     one_exp = []
@@ -175,7 +176,7 @@ def test_noise() -> None:
         .measure_all()
     )
 
-    b.compile_circuit(c2)
+    c2 = b.get_compiled_circuit(c2)
     shots = b.get_shots(c2, 10, seed=5)
     assert shots.shape == (10, 4)
 
@@ -205,7 +206,7 @@ def test_process_characterisation_no_noise_model() -> None:
     assert back.characterisation is None
 
     c = Circuit(4).CX(0, 1).H(2).CX(2, 1).H(3).CX(0, 3).H(1).X(0)
-    back.compile_circuit(c)
+    c = back.get_compiled_circuit(c)
     assert back.valid_circuit(c)
 
 
@@ -226,7 +227,7 @@ def test_process_characterisation_incomplete_noise_model() -> None:
     back = AerBackend(my_noise_model)
 
     c = Circuit(4).CX(0, 1).H(2).CX(2, 1).H(3).CX(0, 3).H(1).X(0).measure_all()
-    back.compile_circuit(c)
+    c = back.get_compiled_circuit(c)
     assert back.valid_circuit(c)
 
     arch = back.backend_info.architecture
@@ -264,7 +265,7 @@ def test_circuit_compilation_complete_noise_model() -> None:
     back = AerBackend(my_noise_model)
 
     c = Circuit(4).CX(0, 1).H(2).CX(2, 1).H(3).CX(0, 3).H(1).X(0).measure_all()
-    back.compile_circuit(c)
+    c = back.get_compiled_circuit(c)
     assert back.valid_circuit(c)
 
 
@@ -300,29 +301,67 @@ def test_process_characterisation_complete_noise_model() -> None:
     back = AerBackend(my_noise_model)
     char = cast(Dict[str, Any], back.characterisation)
 
-    node_errors = cast(Dict[Node, dict], char.get("NodeErrors", {}))
-    link_errors = cast(Dict[Tuple[Node, Node], dict], char.get("EdgeErrors", {}))
+    node_errors = cast(Dict, back.backend_info.all_node_gate_errors)
+    link_errors = cast(Dict, back.backend_info.all_edge_gate_errors)
     arch = back.backend_info.architecture
 
-    assert char["GenericTwoQubitQErrors"][(0, 1)][0][1][0] == 0.0375
-    assert char["GenericTwoQubitQErrors"][(0, 1)][0][1][15] == 0.4375
-    assert char["GenericOneQubitQErrors"][0][0][1][0] == 0.125
-    assert char["GenericOneQubitQErrors"][0][0][1][3] == 0.625
-    assert char["GenericOneQubitQErrors"][0][1][1][0] == 0.35
-    assert char["GenericOneQubitQErrors"][0][1][1][1] == 0.65
-    assert char["GenericOneQubitQErrors"][0][2][1][0] == 0.35
-    assert char["GenericOneQubitQErrors"][0][2][1][1] == 0.65
+    gqe2 = cast(Dict, char["GenericTwoQubitQErrors"])
+    gqe1 = cast(Dict, char["GenericOneQubitQErrors"])
+    assert gqe2[(0, 1)][0][1][0] == 0.0375
+    assert gqe2[(0, 1)][0][1][15] == 0.4375
+    assert gqe1[0][0][1][0] == 0.125
+    assert gqe1[0][0][1][3] == 0.625
+    assert gqe1[0][1][1][0] == 0.35
+    assert gqe1[0][1][1][1] == 0.65
+    assert gqe1[0][2][1][0] == 0.35
+    assert gqe1[0][2][1][1] == 0.65
     assert node_errors[arch.nodes[0]][OpType.U3] == 0.375
     assert link_errors[(arch.nodes[0], arch.nodes[1])][OpType.CX] == 0.5625
     assert link_errors[(arch.nodes[1], arch.nodes[0])][OpType.CX] == 0.80859375
-    assert char["ReadoutErrors"][arch.nodes[0]] == [[0.8, 0.2], [0.2, 0.8]]
-    assert char["ReadoutErrors"][arch.nodes[1]] == [[0.7, 0.3], [0.3, 0.7]]
+    readout_errors = cast(Dict, back.backend_info.all_readout_errors)
+    assert readout_errors[arch.nodes[0]] == [
+        [0.8, 0.2],
+        [0.2, 0.8],
+    ]
+    assert readout_errors[arch.nodes[1]] == [
+        [0.7, 0.3],
+        [0.3, 0.7],
+    ]
+
+
+def test_process_model() -> None:
+    noise_model = NoiseModel()
+    # add readout error to qubits 0, 1, 2
+    error_ro = ReadoutError([[0.8, 0.2], [0.2, 0.8]])
+    for i in range(3):
+        noise_model.add_readout_error(error_ro, [i])
+    # add depolarizing error to qubits 3, 4, 5
+    error_dp_sq = depolarizing_error(0.5, 1)
+    for i in range(3, 6):
+        noise_model.add_quantum_error(error_dp_sq, ["u3"], [i])
+    error_dp_mq = depolarizing_error(0.6, 2)
+    # add coupling errors
+    noise_model.add_quantum_error(error_dp_mq, ["cx"], [0, 7])
+    noise_model.add_quantum_error(error_dp_mq, ["cx"], [1, 2])
+    noise_model.add_quantum_error(error_dp_mq, ["cx"], [8, 9])
+
+    # check basic information has been captured
+    b = AerBackend(noise_model)
+    nodes = b.backend_info.architecture.nodes
+    assert len(nodes) == 9
+    assert "characterisation" in b.backend_info.misc
+    assert "GenericOneQubitQErrors" in b.backend_info.misc["characterisation"]
+    assert "GenericTwoQubitQErrors" in b.backend_info.misc["characterisation"]
+    node_gate_errors = cast(Dict, b.backend_info.all_node_gate_errors)
+    assert nodes[3] in node_gate_errors
+    edge_gate_errors = cast(Dict, b.backend_info.all_edge_gate_errors)
+    assert (nodes[7], nodes[8]) in edge_gate_errors
 
 
 def test_cancellation_aer() -> None:
     b = AerBackend()
     c = circuit_gen(True)
-    b.compile_circuit(c)
+    c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, 10)
     b.cancel(h)
     print(b.circuit_status(h))
@@ -332,7 +371,7 @@ def test_cancellation_aer() -> None:
 def test_cancellation_ibmq(santiago_backend: IBMQBackend) -> None:
     b = santiago_backend
     c = circuit_gen(True)
-    b.compile_circuit(c)
+    c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, 10)
     b.cancel(h)
     print(b.circuit_status(h))
@@ -347,7 +386,7 @@ def test_machine_debug(santiago_backend: IBMQBackend) -> None:
         with pytest.raises(CircuitNotValidError) as errorinfo:
             handles = backend.process_circuits([c, c.copy()], n_shots=2)
         assert "in submitted does not satisfy GateSetPredicate" in str(errorinfo.value)
-        backend.compile_circuit(c)
+        c = backend.get_compiled_circuit(c)
         handles = backend.process_circuits([c, c.copy()], n_shots=4)
         from pytket.extensions.qiskit.backends.ibm import _DEBUG_HANDLE_PREFIX
 
@@ -383,8 +422,7 @@ def test_nshots_batching(santiago_backend: IBMQBackend) -> None:
         c4 = Circuit(2, 2).Rx(0.5, 0).CX(0, 1).CX(1, 0).measure_all()
         cs = [c1, c2, c3, c4]
         n_shots = [10, 12, 10, 13]
-        for c in cs:
-            backend.compile_circuit(c)
+        cs = backend.get_compiled_circuits(cs)
         handles = backend.process_circuits(cs, n_shots=n_shots)
 
         from pytket.extensions.qiskit.backends.ibm import _DEBUG_HANDLE_PREFIX
@@ -552,8 +590,7 @@ def test_swaps_basisorder() -> None:
     c1 = cu.circuit
     assert c1.n_gates_of_type(OpType.CX) == 2
 
-    b.compile_circuit(c)
-    b.compile_circuit(c1)
+    c, c1 = b.get_compiled_circuits([c, c1])
 
     handles = b.process_circuits([c, c1])
     s_ilo = b.get_state(c1, basis=BasisOrder.ilo)
@@ -592,7 +629,7 @@ def test_pauli() -> None:
     for b in [AerBackend(), AerStateBackend()]:
         c = Circuit(2)
         c.Rz(0.5, 0)
-        b.compile_circuit(c)
+        c = b.get_compiled_circuit(c)
         zi = QubitPauliString(Qubit(0), Pauli.Z)
         assert cmath.isclose(get_pauli_expectation_value(c, zi, b), 1)
         c.X(0)
@@ -608,6 +645,18 @@ def test_operator() -> None:
         assert cmath.isclose(get_operator_expectation_value(c, zz, b), 1.0)
         c.X(0)
         assert cmath.isclose(get_operator_expectation_value(c, zz, b), -1.0)
+
+
+# TKET-1432 this was either too slow or consumed too much memory when bugged
+@pytest.mark.timeout(10)
+def test_expectation_bug() -> None:
+    backend = AerStateBackend()
+    # backend.compile_circuit(circuit)
+    circuit = Circuit(16)
+    with open("big_hamiltonian.pickle", "rb") as f:
+        hamiltonian = pickle.load(f)
+    exp = backend.get_operator_expectation_value(circuit, hamiltonian)
+    assert np.isclose(exp, 1.4325392)
 
 
 def test_aer_result_handle() -> None:
@@ -659,7 +708,7 @@ def test_aerstate_result_handle() -> None:
 def test_cache() -> None:
     b = AerBackend()
     c = circuit_gen()
-    b.compile_circuit(c)
+    c = b.get_compiled_circuit(c)
     h = b.process_circuits([c], 2)[0]
     b.get_result(h).get_shots()
     assert h in b._cache
@@ -683,7 +732,7 @@ def test_mixed_circuit() -> None:
     c.X(qr[1], condition=reg_eq(ar, 0))
     c.Measure(qr[1], br[0])
     backend = AerBackend()
-    backend.compile_circuit(c)
+    c = backend.get_compiled_circuit(c)
     counts = backend.get_counts(c, 1024)
     for key in counts.keys():
         assert key in {(0, 1), (1, 0)}
@@ -745,12 +794,12 @@ def test_ibmq_emulator() -> None:
             assert all(pred.verify(c_cop) for pred in bac.required_predicates)
 
         c_cop_2 = c.copy()
-        b_aer.compile_circuit(c_cop_2, ol)
+        c_cop_2 = b_aer.get_compiled_circuit(c_cop_2, ol)
         if ol == 0:
             assert not all(pred.verify(c_cop_2) for pred in b_emu.required_predicates)
 
     circ = Circuit(2, 2).H(0).CX(0, 1).measure_all()
-    b_emu.compile_circuit(circ)
+    circ = b_emu.get_compiled_circuit(circ)
     b_noi = AerBackend(noise_model=b_emu._noise_model)
     emu_shots = b_emu.get_shots(circ, 10, seed=10)
     aer_shots = b_noi.get_shots(circ, 10, seed=10)
@@ -841,7 +890,7 @@ def test_ibmq_mid_measure(santiago_backend: IBMQBackend) -> None:
     b = santiago_backend
     ps = b.default_compilation_pass(0)
     ps.apply(c)
-    # b.compile_circuit(c)
+    # c = b.get_compiled_circuit(c)
     assert not NoMidMeasurePredicate().verify(c)
     assert b.valid_circuit(c)
 
@@ -853,7 +902,7 @@ def test_compile_x(santiago_backend: IBMQBackend) -> None:
     c = Circuit(1).X(0)
     for ol in range(3):
         c1 = c.copy()
-        b.compile_circuit(c1, optimisation_level=ol)
+        c1 = b.get_compiled_circuit(c1, optimisation_level=ol)
         assert c1.n_gates == 1
 
 
@@ -963,7 +1012,7 @@ def test_postprocess(santiago_backend: IBMQBackend) -> None:
     assert b.supports_contextual_optimisation
     c = Circuit(2, 2)
     c.SX(0).SX(1).CX(0, 1).measure_all()
-    b.compile_circuit(c)
+    c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, n_shots=10, postprocess=True)
     ppcirc = Circuit.from_dict(json.loads(cast(str, h[2])))
     ppcmds = ppcirc.get_commands()
@@ -978,7 +1027,7 @@ def test_postprocess_emu() -> None:
     assert b.supports_contextual_optimisation
     c = Circuit(2, 2)
     c.SX(0).SX(1).CX(0, 1).measure_all()
-    b.compile_circuit(c)
+    c = b.get_compiled_circuit(c)
     h = b.process_circuit(c, n_shots=10, postprocess=True)
     ppcirc = Circuit.from_dict(json.loads(cast(str, h[2])))
     ppcmds = ppcirc.get_commands()