+++ modules/pytket-qsharp/docs/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 ~~~~~~~~~
 
+0.15.1 (unreleased)
+-------------------
+
+* Updated qsharp version requirement to 0.18.
+
+0.15.0 (July 2021)
+------------------
+
+* Updated pytket version requirement to 0.13.
+
 0.14.0 (June 2021)
 ------------------